Mixins conflicting with Angelica now only load when it is not present
