+++ src/main/java/su/sergiusonesimus/metaworlds/zmixin/Mixins.java
@@ -87,12 +87,18 @@ public enum Mixins {
             .setPhase(Phase.LATE)
             .addMixinClasses("beddium.MixinRenderGlobal")),
 
-    ANGELICA_COMPAT(new Builder(
-        "Reenable vanilla rendering disabled by Angelica for subworlds. Not the best fix, but a fix nevertheless")
-            .addTargetedMod(TargetedMod.ANGELICA)
-            .setSide(Side.CLIENT)
-            .setPhase(Phase.LATE)
-            .addMixinClasses(addPrefix("angelica.", "MixinRenderGlobal", "MixinEffectRenderer", "MixinClientProxy"))),
+    NO_ANGELICA(new Builder("").addTargetedMod(TargetedMod.VANILLA)
+        .addExcludedMod(TargetedMod.ANGELICA)
+        .setSide(Side.CLIENT)
+        .setPhase(Phase.EARLY)
+        .addMixinClasses("minecraft.client.renderer.MixinRenderGlobalVanilla")),
+
+    // ANGELICA_COMPAT(new Builder(
+    // "Reenable vanilla rendering disabled by Angelica for subworlds. Not the best fix, but a fix nevertheless")
+    // .addTargetedMod(TargetedMod.ANGELICA)
+    // .setSide(Side.CLIENT)
+    // .setPhase(Phase.EARLY)
+    // .addMixinClasses(addPrefix("angelica.", "MixinRenderGlobal", "MixinEffectRenderer", "MixinClientProxy"))),
 
     HARDCORE_ENDER_EXPANSION_COMPAT(new Builder("Disable generation of additional data for player proxies")
         .addTargetedMod(TargetedMod.HARDCORE_ENDER_EXPANSION)
